feat(cap_tools.models): *_to_dict() now return MultiDict, *_from_dict() now accept Mapping
This change adds support for repeated occurances of the same "key" in Info.parameters, Info.event_codes and Area.geocode when using the *_to_dict() and *_from_dict() methods. While this was classified as a non-breaking change, there is a slight change your code might still break when you checked for the exact type returned (e.g. used isinstance()) or when you relied on the overwriting behavior of dicts for consecutive occurances of the same key, i.e. dict would (only) retain the last occurance while MultiDict.__getitem__() returns the first.